app/requirements.txt
python-telegram-bot==21.6
routeros-api==0.21.0
python-dotenv==1.0.1
